Refactor: 정렬 컴포넌트 다시 분리
