+++ src/pages/SeriesPostList/PostList/Sort/index.jsx
@@ -1,40 +1,64 @@
+import { useState } from 'react';
+import { IoChevronUp, IoChevronDown } from 'react-icons/io5';
 import styled from 'styled-components';
 
-const SortButton = ({ icon, text, onClick }) => {
+const Sort = () => {
+  const [isSort, setIsSort] = useState(true);
+
+  const onSort = () => {
+    setIsSort(!isSort);
+  };
+
   return (
-    <SortButtonContainer text={text} onClick={onClick}>
-      {icon}
-      <span>{text}</span>
-    </SortButtonContainer>
+    <SortContainer>
+      <button onClick={onSort}>
+        {isSort ? (
+          <>
+            <IoChevronUp /> <span>오름차순</span>
+          </>
+        ) : (
+          <>
+            <IoChevronDown /> <span>내림차순</span>
+          </>
+        )}
+      </button>
+    </SortContainer>
   );
 };
 
-const SortButtonContainer = styled.button`
-  display: inline-flex;
-  -webkit-box-align: center;
-  align-items: center;
-  height: 2rem;
-  padding-left: 0.5rem;
-  padding-right: 0.75rem;
+const SortContainer = styled.div`
+  display: flex;
+  -webkit-box-pack: end;
+  justify-content: flex-end;
+  margin-top: 1rem;
 
-  background: var(--bg-element2);
-  border-radius: 4px;
-  border: none;
-  outline: none;
-  cursor: pointer;
+  button {
+    display: inline-flex;
+    -webkit-box-align: center;
+    align-items: center;
+    height: 2rem;
+    padding-left: 0.5rem;
+    padding-right: 0.75rem;
 
-  svg {
-    color: var(--primary2);
-    font-size: 1.5rem;
-    transition: all 0.125s ease-in 0s;
-  }
+    background: var(--bg-element2);
+    border-radius: 4px;
+    border: none;
+    outline: none;
+    cursor: pointer;
 
-  span {
-    margin-left: 0.25rem;
-    font-size: 1rem;
-    color: var(--text1);
-    line-height: 1;
+    svg {
+      color: var(--primary2);
+      font-size: 1.5rem;
+      transition: all 0.125s ease-in 0s;
+    }
+
+    span {
+      margin-left: 0.25rem;
+      font-size: 1rem;
+      color: var(--text1);
+      line-height: 1;
+    }
   }
 `;
 
-export default SortButton;
+export default Sort;
